everything.xml: Remove zephyr app builder
This component is not really adding much value at this point.
Zephyr's build targetting simulated targets is very straight forward
at this point and nobody is expected to be using this anymore.
In any case, as is, this makefile was more of an example than
a functional component.
Let's just remove it from the manifest.
If somebody still wants it they can fetch it by hand.

Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>