everything.xml: Remove ext_NRF52_hw_models component
This component was deprecated in favour of the
ext_NRF_hw_models in 2023/08.

Although this old component could still be used,
the mayority of users have transitioned to its replacement.
So keeping it in this manifest is just crud for
the mayority.
Moreover as this component requires special setup to build
(environment variables set), it was confusing for users.

Let's remove it.

Those who still want to use it can do so by fetching
it by hand or with their own manifest.

Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>